pyproject.toml
[project]
name = "azure-functions-python-cookbook"
description = "Practical recipes and examples for Azure Functions Python"
keywords = ["azure-functions", "azure", "python", "v2", "cookbook", "recipes", "examples", "serverless"]
authors = [{ name = "Yeongseon Choe", email = "yeongseon.choe@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10,<3.15"
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = []

[project.optional-dependencies]
dev = [
    "azure-functions",
    "azure-functions-durable",
    "azure-functions-openapi",
    "azure-identity",
    "azure-storage-blob",
    "azure-storage-queue",
    "bandit==1.9.4",
    "build",
    "coverage",
    "git-cliff",
    "hatch",
    "mypy==1.19.1",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "requests",
    "ruff==0.15.6",
    "twine",
]
docs = [
    "mkdocs<2.0",
    "mkdocs-material<10.0",
    "mkdocstrings[python]<2.0",
]

[project.urls]
Homepage = "https://yeongseon.github.io/azure-functions-python-cookbook/"
Documentation = "https://yeongseon.github.io/azure-functions-python-cookbook/"
Repository = "https://github.com/yeongseon/azure-functions-python-cookbook"
Issues = "https://github.com/yeongseon/azure-functions-python-cookbook/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "/recipes",
    "/examples",
    "/README.md",
    "/Makefile",
    "/pyproject.toml",
]

[tool.hatch.version]
path = "src/azure_functions_python_cookbook/__init__.py"

# NOTE: This project is a content-first cookbook (recipes + examples), not a runtime library.
# The publish config below exists for potential future recipe-index or utility distribution.
# There is no published PyPI package today. See docs/api.md for details.
[tool.hatch.publish.index]
default = "pypi"

[tool.hatch.publish.index.repos.pypi]
url = "https://upload.pypi.org/legacy/"

[tool.hatch.publish.index.repos.test]
url = "https://test.pypi.org/legacy/"

[tool.hatch.envs.default]
python = "3.10"
features = ["dev", "docs"]

[tool.hatch.envs.default.scripts]
format = "ruff format src tests"
style = "ruff check src tests"
typecheck = "mypy src tests"
lint = "hatch run style && hatch run typecheck"
test = "pytest -v tests"
cov = "pytest --cov-report=term-missing --cov-report=html --cov-report=xml tests/"
docs = "mkdocs serve"
precommit = "pre-commit run --all-files"
precommit-install = "pre-commit install"
security = "python -m bandit -r src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["azure_functions_python_cookbook"]
force-sort-within-sections = true
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
show_error_codes = true
pretty = true
exclude = ["tests/e2e/"]

[tool.pytest.ini_options]
addopts = "--cov=src/azure_functions_python_cookbook --cov-report=xml --cov-report=term-missing -ra -q -m 'not (e2e or smoke)'"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "e2e: End-to-end tests requiring Azurite and func host",
    "smoke: Module-import smoke tests for non-emulatable triggers",
]

[tool.coverage.run]
branch = true
source = ["src/azure_functions_python_cookbook"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 90

[tool.bandit]
exclude_dirs = ["tests"]
